Gross hack so embedded vids from Super Markdown work on personal blogs
Basically, the video info is actually saved physically inside the tag
as base64 when the video is embedded by the editor, and then xkit_patches
uses that base64 to restore the actual video. This is about the only
approach I could find that Tumblr didn't ruthlessly strip out of the HTML.

It's in xkit_patches because it should apply even if you don't use Super
Markdown Editor yourself (when you look at other people's posts). I'd
prefer it to be actually part of Tumblr, but we aren't gods. Also, if we
had that level of control I'd just fix Tumblr straight up so we don't
actually need disgusting hacks like this.