Don't hide the grey boxes under embedded videos, it's Tumblr's problem
Tumblr's handling of its own embed code inside HTML or Markdown posts is
broken such that it generates grey "invalid tag" boxes even though the
dash can in fact (and does) handle the embedded video. But that's the site
being broken, and we should be consistent with non-XKit users so don't
patch it in XKit.