blacklist.js - preserve href links
Spam Tumblrs can often be identified by the presence of an external link - often to http://ift.tt or http://smarturl.it.  However, you can't currently blacklist those URLs if the link text isn't the URL itself.  That is, <a href="http://spamlink.com">Click here</a> isn't caught with a blacklist of "http://spamlink.com*".  This patch fixes that problem.