Replace run checks with checking for tumblr page
For some reason, Tag Tracking+ had a check built in by Atesh to not make
it run on certain pages. It only ran on the dashboard, which is not
really wanted anymore due to Tag Tracking+'s extended functionality.
The checks have been replaced with Interface's is_tumblr_page to prevent
it from running on blogs.