Update go_to_dash.js
Wrapped "View" in a span to give the go-to-dash same styling as other post buttons;
I think the original extension stripped out the dashboard button text to make room for the view button, but tumblr must've changed something because the old code wasn't doing anything. This removes the dashboard button label directly and tweaks margins with CSS so the dashboard button goes back to being just the home symbol.