Prioritised current Retags behaviour over old look
Retags seems to insert tags after the Activity+ timestamp rather than before, as was the case before the tumblr update. The goal of this tweak was to give the activity feed its exact old look back but the way Retags and Activity+ timestamps work now, that is impossible, so this is the next best thing.