User Menus+: fix handling of mentions (#1331)
'Mentions' in Tumblr posts link not to the usual person.tumblr.com
link, but to a link on a site called tmblr.co with a long string of
characters to identify a blog (presumably in a more permanent way). User
Menus+ was not able to retrieve 'tumbelog popover' data from these kinds
of links.

Now, an additional test for tmblr.co links is performed, and popover
data can be correctly retrieved for mentions.

Additionally there is some cleanup: all logging messages go the XKit
console rather than a haphazard mixture of the XKit console and browser
console.