AutoSync Introduction to XCloud
Minor stuff: version bump, details typo correction
Major stuff:
Adds a preference to enable AutoSync, and a basic explanation to what it is and how it works
Modifies start_fetch and process_restore to take the argument "auto"; if true, the received data is treated differently
While it was impossible to attempt synchronisation without logging in, enabling AutoSync means the login screen can be bypassed. So, if they're not logged in, AutoSync complains because it's the only way the user could have done that
If the internal extensions exist in the cloud, they get put in the list first, so if auto is true the restoration loop will go through xkit_main and xkit_patches (and ignore them) before getting to xkit_preferences to check the timestamp. Now that we have xkit_preferences, the local timestamp is saved before it's overwritten by the XCloud data so the two can be compared
If XCloud's timestamp is older the newer timestamp gets written back in and the upload process starts
If XCloud's timestamp is newer, we can just leave it there, and everything else can be restored as per usual (albeit skipping xkit_main and xkit_patches, but those were completely skipped before anyway)
If the timestamps match, do nothing but display a happy window with some reassuring numbers (both timestamps) that closes itself after 2 seconds
And of course, to allow the whole thing to work, the uploader doesn't skip the internal extensions anymore, although xkit_main and xkit_patches will only be readable from the cloud manually