Use peepr data to replace original_url
I couldn't find an original_url href to use so this instead uses the
peepr data to get the post id and blog name. It checks this post id
against the notificationblock-provided blacklist. Next, it uses both to
create a fake original_url for notificationblock's block function to
consume later.

Fixes #1345.