J/K Across Pages: smarter navigation
Users who "J" to the next page then scroll past the first post are no longer yanked back to the first post if the extension takes too long to initialise.
Replaced window.hash with history.replaceState as browsers would have jumped back to the top of the page and discarded the user's scroll position otherwise.