Mass Re-organisation and Cleaning Up
Ok! It's taken a while, but thankfully I'm sick and on holidays so I didn't rly have much else to do BUT this XD This should have the same end result as the first commit, but this time it should make a lot more sense as to what's happening here ^__^

Also added a few more neat graphics while I was it!