ensure post existence when fetching timestamp
The peepr API, when asked for a post that has been deleted, does not explicitly fail; instead, it returns a "Not Found" message alongside a response for the blog as if no post-related parameters had been inputted.
This caused a bug in reblog timestamps where a deleted post would be timestamped with the time of the blog's most recent post.