[Tweaks] More carefully target notification badges
Some notification badges are when getting activity on a sideblog, or when receiving a message on an account other than your own (in the blog dropdown of the Messaging pop-up). These badges do not look good when recoloured, which is especially noticable if they are white in colour. The new selectors only target notification badges that are in the header, and not ones that are in descendant drop-downs or those on minified messaging windows.